+++ IT/src/main/java/module-info.java
@@ -10,9 +10,9 @@ open module com.microsoft.gctoolkit.integration {
     requires com.microsoft.gctoolkit.parser;
     requires com.microsoft.gctoolkit.vertx;
     requires java.logging;
-    requires org.junit.jupiter.api;
 
-    exports com.microsoft.gctoolkit.integration;
+//    exports com.microsoft.gctoolkit.integration.aggregation;
+//    exports com.microsoft.gctoolkit.integration.collections;
 
     exports com.microsoft.gctoolkit.integration.aggregation to
             com.microsoft.gctoolkit.vertx;
