+++ api/src/main/java/com/microsoft/gctoolkit/event/MalformedEvent.java
@@ -1,4 +1,7 @@
 package com.microsoft.gctoolkit.event;
 
-public class MalformedEvent {
+public class MalformedEvent extends Exception {
+    public MalformedEvent(String message) {
+        super(message);
+    }
 }
