+++ vertx/src/main/java/com/microsoft/gctoolkit/vertx/jvm/PreUnifiedJavaVirtualMachine.java
@@ -31,4 +31,14 @@ public class PreUnifiedJavaVirtualMachine extends AbstractJavaVirtualMachine {
         return new GCToolkitVertxParametersForPreUnifiedLogs(registeredAggregations, diary);
     }
 
+    @Override
+    public boolean isZGC() {
+        return false;
+    }
+
+    @Override
+    public boolean isShenandoah() {
+        return false;
+    }
+
 }
